CHORE: restructure Arcus.Security.Providers.AzureKeyVault
Restructure the AzureKeyVault providers by:
- Including the `KeyVaultSecretProvider`
- Placing the `IKeyVaultConfig...` and `IKeyVaultAuth...` in the
repectable `Config...` and `Auth...` folders i.o. sep. `Interfaces`
folder.
- Update the documentation to remove the
`Arcus.Security.Secrets.AzureKeyVault` NuGet package

Relates to #100
