+++ src/Arcus.Security.Providers.AzureKeyVault/Authentication/IKeyVaultAuthentication.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.Azure.KeyVault;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
+using Microsoft.Azure.KeyVault;
 
-namespace Arcus.Security.Providers.AzureKeyVault.Authentication.Interfaces
+namespace Arcus.Security.Providers.AzureKeyVault.Authentication
 {
     /// <summary>
     ///     Authentication provider for Azure Key Vault
