+++ src/Arcus.Security.Providers.AzureKeyVault/KeyVaultSecretProvider.cs
@@ -2,8 +2,8 @@
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using Arcus.Security.Providers.AzureKeyVault.Authentication.Interfaces;
-using Arcus.Security.Providers.AzureKeyVault.Configuration.Interfaces;
+using Arcus.Security.Providers.AzureKeyVault.Authentication;
+using Arcus.Security.Providers.AzureKeyVault.Configuration;
 using Arcus.Security.Secrets.Core.Exceptions;
 using Arcus.Security.Secrets.Core.Interfaces;
 using Arcus.Security.Secrets.Core.Models;
@@ -12,7 +12,7 @@ using Microsoft.Azure.KeyVault;
 using Microsoft.Azure.KeyVault.Models;
 using Polly;
 
-namespace Arcus.Security.Secrets.AzureKeyVault
+namespace Arcus.Security.Providers.AzureKeyVault
 {
     /// <summary>
     ///     Secret key provider that connects to Azure Key Vault
