--- src/Arcus.Security.Secrets.AzureKeyVault/Arcus.Security.Secrets.AzureKeyVault.csproj
@@ -1,28 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <Authors>Arcus</Authors>
-    <Description>Provides support for Azure Key Vault</Description>
-    <Copyright>Copyright (c) Arcus</Copyright>
-    <PackageLicenseUrl>https://github.com/arcus-azure/arcus.security/blob/master/LICENSE</PackageLicenseUrl>
-    <PackageProjectUrl>https://github.com/arcus-azure/arcus.security</PackageProjectUrl>
-    <PackageIconUrl>https://raw.githubusercontent.com/arcus-azure/arcus/master/media/arcus.png</PackageIconUrl>
-    <RepositoryUrl>https://github.com/arcus-azure/arcus.security</RepositoryUrl>
-    <RepositoryType>Git</RepositoryType>
-    <PackageTags>Azure;Key Vault;Security</PackageTags>
-    <PackageId>Arcus.Security.Secrets.AzureKeyVault</PackageId>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Polly-Signed" Version="5.9.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Arcus.Security.Providers.AzureKeyVault\Arcus.Security.Providers.AzureKeyVault.csproj" />
-    <ProjectReference Include="..\Arcus.Security.Secrets.Core\Arcus.Security.Secrets.Core.csproj" />
-  </ItemGroup>
-
-</Project>
