+++ src/Arcus.Security.Tests.Unit/KeyVault/KeyVaultSecretProviderTests.cs
@@ -1,5 +1,4 @@
 ﻿using Arcus.Security.Providers.AzureKeyVault.Configuration;
-using Arcus.Security.Secrets.AzureKeyVault;
 using Arcus.Security.Secrets.Core.Models;
 using Arcus.Security.Tests.Unit.KeyVault.Doubles;
 using Microsoft.Azure.KeyVault.Models;
@@ -8,6 +7,7 @@ using System;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Arcus.Security.Providers.AzureKeyVault;
 using Xunit;
 
 namespace Arcus.Security.Tests.Unit.KeyVault
