+++ src/Arcus.Security.sln
@@ -17,8 +17,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Secrets", "Secrets", "{26FD
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Providers", "Providers", "{DD66B1E8-3676-4C13-8C37-AEA80E1C21B7}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Secrets.AzureKeyVault", "Arcus.Security.Secrets.AzureKeyVault\Arcus.Security.Secrets.AzureKeyVault.csproj", "{473728CE-777E-4FC3-8164-6B603088C877}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{7E59A380-48DB-443B-9B47-05A4A175B8F3}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Core", "Arcus.Security.Core\Arcus.Security.Core.csproj", "{EC4CBB2C-9C38-4D5C-B4C4-602F0756A61F}"
@@ -47,10 +45,6 @@ Global
 		{DD03DFBB-7D61-4FC9-B2D0-1BB800EA1ED9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DD03DFBB-7D61-4FC9-B2D0-1BB800EA1ED9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DD03DFBB-7D61-4FC9-B2D0-1BB800EA1ED9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{473728CE-777E-4FC3-8164-6B603088C877}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{473728CE-777E-4FC3-8164-6B603088C877}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{473728CE-777E-4FC3-8164-6B603088C877}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{473728CE-777E-4FC3-8164-6B603088C877}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EC4CBB2C-9C38-4D5C-B4C4-602F0756A61F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EC4CBB2C-9C38-4D5C-B4C4-602F0756A61F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EC4CBB2C-9C38-4D5C-B4C4-602F0756A61F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -68,7 +62,6 @@ Global
 		{DACBC3C6-2894-4961-B327-746BC29229CC} = {C42FDAC3-5A31-4F39-BD92-C57C2CDAB2D2}
 		{D185D73E-D617-4846-9CAB-12884C1266C5} = {C42FDAC3-5A31-4F39-BD92-C57C2CDAB2D2}
 		{DD03DFBB-7D61-4FC9-B2D0-1BB800EA1ED9} = {DD66B1E8-3676-4C13-8C37-AEA80E1C21B7}
-		{473728CE-777E-4FC3-8164-6B603088C877} = {26FD7AFD-F12E-4F1F-965E-379065340D47}
 		{EC4CBB2C-9C38-4D5C-B4C4-602F0756A61F} = {7E59A380-48DB-443B-9B47-05A4A175B8F3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
